# backend/doc/swagger.yaml
swagger: '2.0'
info:
  title: Appli Atelier
  description: Backend de l'appli Atelier de Maintenance. Toutes les données sont encapsulées dans une réponse conforme au standard "JSend" (https://labs.omniti.com/labs/jsend)
  version: 0.0.1
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /actions:
    get:
      summary: Récupère les actions existantes
      description: Récupère la liste des actions existantes
      parameters:
        - in: query
          name: equipment
          type: integer
          description: ID de l'équipement pour filtrer les actions
        - in: query
          name: search
          type: string
          description: début du nom d'une action pour l'autocomplete
      responses:
        '200':
          description: Récupération réussie
          schema:
            type: object
            required:
              - status
            properties:
              status:
                type: string
                description: statut selon le standard Jsend
                enum: [
                  "success",
                  "fail",
                  "error"
                ]
              code:
                type: integer
                description: code numérique en cas d'erreur selon le standard Jsend
              message:
                type: string
                description: message en cas d'erreur selon le standard Jsend
              data:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: l'id de l'action
                    name:
                      type: string
                      description: nom de l'action
        '401':
          description: Not authenticated
    post:
      summary: Créer une action
      description: Créer une nouvelle action
      parameters:
        - in: body
          name: action
          description: nouvelle action à créer
          schema:
            type: object
            required:
              - equipmentId
              - name
            properties:
              equipmentId:
                type: integer
                description: id de l'équipement concerné par l'action  
              name:
                type: string
                description: le nom de l'action
      responses:
        '200':
          description: Création réussie
          schema:
            type: object
            required:
              - status
            properties:
              status:
                type: string
                description: statut selon le standard Jsend
                enum: [
                  "success",
                  "fail",
                  "error"
                ]
              code:
                type: integer
                description: code numérique en cas d'erreur selon le standard Jsend
              message:
                type: string
                description: message en cas d'erreur selon le standard Jsend
              data:
                type: object
                properties:
                  id:
                    type: integer
                    description: l'id de l'action
                  name:
                    type: string
                    description: nom de l'action
        '400':
          description: Action existe déjà
        '401':
          description: Erreur d'authentification
...